+++ example/a.ts
@@ -2,9 +2,16 @@ import { Elysia } from '../src'
 import { req } from '../test/utils'
 
 const app = new Elysia({ precompile: true })
-    .get('/hello', ({ query: { name } }) => name)
-    .listen(3000)
+	.get('/hello', ({ cookie: { name } }) => {
+		name.value = 'SaltyAom'
+		name.partitioned = false
 
-console.log(app.routes[0].composed.toString())
+		return `Hello, ${name}!`
+	})
+	.listen(3000)
 
-app.handle(req('/hello?name=hello+123')).then(x => x.text()).then(x => console.log({ x }))
+// console.log(app.routes[0].composed.toString())
+
+app.handle(req('/hello?name=hello+123'))
+	.then((x) => x.text())
+	.then((x) => console.log({ x }))