+++ src/cookies.ts
@@ -68,6 +68,17 @@ export interface CookieOptions {
 	 */
 	priority?: 'low' | 'medium' | 'high' | undefined
 	/**
+	 * Specifies the `boolean` value for the [`Partitioned` `Set-Cookie`](rfc-cutler-httpbis-partitioned-cookies)
+	 * attribute. When truthy, the `Partitioned` attribute is set, otherwise it is not. By default, the
+	 * `Partitioned` attribute is not set.
+	 *
+	 * **note** This is an attribute that has not yet been fully standardized, and may change in the future.
+	 * This also means many clients may ignore this attribute until they understand it.
+	 *
+	 * More information about can be found in [the proposal](https://github.com/privacycg/CHIPS)
+	 */
+	partitioned?: boolean | undefined
+	/**
 	 * Specifies the boolean or string to be the value for the {@link https://tools.ietf.org/html/draft-ietf-httpbis-rfc6265bis-03#section-4.1.2.7|`SameSite` `Set-Cookie` attribute}.
 	 *
 	 * - `true` will set the `SameSite` attribute to `Strict` for strict same
@@ -207,6 +218,14 @@ export class Cookie<T> implements ElysiaCookie {
 		this.cookie.priority = priority
 	}
 
+	get partitioned() {
+		return this.cookie.partitioned
+	}
+
+	set partitioned(partitioned) {
+		this.cookie.partitioned = partitioned
+	}
+
 	get secrets() {
 		return this.cookie.secrets
 	}
@@ -228,7 +247,7 @@ export class Cookie<T> implements ElysiaCookie {
 		this.cookie = Object.assign(
 			{
 				...this.initial,
-				value: this.value,
+				value: this.value
 			},
 			typeof config === 'function' ? config(this.cookie) : config
 		)